+++ packages/modules/src/modules/epics/ContributionsEpicCtas.tsx
@@ -3,7 +3,7 @@ import { EpicProps } from '@sdc/shared/types';
 import { ContributionsEpicReminder } from './ContributionsEpicReminder';
 import { ContributionsEpicButtons } from './ContributionsEpicButtons';
 import { defineFetchEmail } from '../shared/helpers/definedFetchEmail';
-import { ChoiceCardSelection } from '../banners/choiceCardsBanner/choiceCards';
+import { ChoiceCardSelection } from '../../hooks/choiceCards';
 
 interface OnReminderOpen {
     buttonCopyAsString: string;
